fix: restore Twilio secure-key import boundary
